everest_app/cms_blog/api/serializers.py
from rest_framework import serializers
from cms_blog.models import Post

class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = '__all__'
        depth = 1